+++ x-pack/plugins/lists/public/exceptions/components/autocomplete/helpers.ts
@@ -7,8 +7,9 @@
 
 import dateMath from '@elastic/datemath';
 import { EuiComboBoxOptionOption } from '@elastic/eui';
+import type { Type } from '@kbn/securitysolution-io-ts-list-types';
 
-import { ListSchema, Type } from '../../../../common';
+import type { ListSchema } from '../../../../common';
 import { IFieldType } from '../../../../../../../src/plugins/data/common';
 
 import {
