+++ x-pack/plugins/lists/public/exceptions/components/builder/helpers.ts
@@ -8,27 +8,29 @@
 import uuid from 'uuid';
 import { addIdToItem, removeIdFromItem } from '@kbn/securitysolution-utils';
 import { validate } from '@kbn/securitysolution-io-ts-utils';
-import { OsTypeArray } from '@kbn/securitysolution-io-ts-list-types';
-
-import { IFieldType, IIndexPattern } from '../../../../../../../src/plugins/data/public';
 import {
-  CreateExceptionListItemSchema,
   EntriesArray,
   Entry,
   EntryNested,
-  ExceptionListItemSchema,
   ExceptionListType,
-  ListSchema,
   NamespaceType,
-  OperatorEnum,
-  OperatorTypeEnum,
-  createExceptionListItemSchema,
+  ListOperatorEnum as OperatorEnum,
+  ListOperatorTypeEnum as OperatorTypeEnum,
+  OsTypeArray,
   entriesList,
   entriesNested,
   entry,
-  exceptionListItemSchema,
   nestedEntryItem,
+} from '@kbn/securitysolution-io-ts-list-types';
+
+import {
+  CreateExceptionListItemSchema,
+  ExceptionListItemSchema,
+  ListSchema,
+  createExceptionListItemSchema,
+  exceptionListItemSchema,
 } from '../../../../common';
+import { IFieldType, IIndexPattern } from '../../../../../../../src/plugins/data/public';
 import {
   EXCEPTION_OPERATORS,
   EXCEPTION_OPERATORS_SANS_LISTS,
@@ -96,7 +98,7 @@ export const filterExceptionItems = (
         return [...acc, item];
       } else if (createExceptionListItemSchema.is(item)) {
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const { meta: _, ...rest } = item;
+        const { meta, ...rest } = item;
         const itemSansMetaId: CreateExceptionListItemSchema = { ...rest, meta: undefined };
         return [...acc, itemSansMetaId];
       } else {
