+++ x-pack/plugins/lists/public/shared_exports.ts
@@ -7,11 +7,8 @@
 
 // Exports to be shared with plugins
 export { withOptionalSignal } from './common/with_optional_signal';
-export { useIsMounted } from './common/hooks/use_is_mounted';
 export { useAsync } from './common/hooks/use_async';
 export { useApi } from './exceptions/hooks/use_api';
-export { usePersistExceptionItem } from './exceptions/hooks/persist_exception_item';
-export { usePersistExceptionList } from './exceptions/hooks/persist_exception_list';
 export { useExceptionListItems } from './exceptions/hooks/use_exception_list_items';
 export { useExceptionLists } from './exceptions/hooks/use_exception_lists';
 export { useFindLists } from './lists/hooks/use_find_lists';
@@ -24,13 +21,18 @@ export { useReadListIndex } from './lists/hooks/use_read_list_index';
 export { useCreateListIndex } from './lists/hooks/use_create_list_index';
 export { useReadListPrivileges } from './lists/hooks/use_read_list_privileges';
 export {
-  addExceptionListItem,
-  updateExceptionListItem,
+  getEntryValue,
+  getExceptionOperatorSelect,
+  getOperatorType,
+  getNewExceptionItem,
+  addIdToEntries,
+} from './exceptions/components/builder/helpers';
+export {
   fetchExceptionListById,
   addExceptionList,
   addEndpointExceptionList,
 } from './exceptions/api';
-export {
+export type {
   ExceptionList,
   ExceptionListFilter,
   ExceptionListIdentifiers,
