+++ x-pack/plugins/security_solution/common/detection_engine/schemas/types/lists.ts
@@ -7,7 +7,7 @@
 
 import * as t from 'io-ts';
 
-import { exceptionListType, namespaceType } from '../../../shared_imports';
+import { exceptionListType, namespaceType } from '@kbn/securitysolution-io-ts-list-types';
 
 import { NonEmptyString } from './non_empty_string';
