+++ x-pack/plugins/security_solution/common/shared_imports.ts
@@ -7,44 +7,14 @@
 
 export {
   ListSchema,
-  CommentsArray,
-  CreateCommentsArray,
-  Comment,
-  CreateComment,
   ExceptionListSchema,
   ExceptionListItemSchema,
   CreateExceptionListSchema,
   CreateExceptionListItemSchema,
   UpdateExceptionListItemSchema,
-  Entry,
-  EntryExists,
-  EntryMatch,
-  EntryMatchAny,
-  EntryMatchWildcard,
-  EntryNested,
-  EntryList,
-  EntriesArray,
-  NamespaceType,
-  Operator,
-  OperatorEnum,
-  OperatorTypeEnum,
-  ExceptionListTypeEnum,
   exceptionListItemSchema,
-  exceptionListType,
-  comment,
   createExceptionListItemSchema,
   listSchema,
-  entry,
-  entriesNested,
-  nestedEntryItem,
-  entriesMatch,
-  entriesMatchAny,
-  entriesMatchWildcard,
-  entriesExists,
-  entriesList,
-  namespaceType,
-  ExceptionListType,
-  Type,
   ENDPOINT_LIST_ID,
   ENDPOINT_TRUSTED_APPS_LIST_ID,
   EXCEPTION_LIST_URL,
@@ -52,8 +22,5 @@ export {
   ENDPOINT_EVENT_FILTERS_LIST_ID,
   ENDPOINT_EVENT_FILTERS_LIST_NAME,
   ENDPOINT_EVENT_FILTERS_LIST_DESCRIPTION,
-  osType,
-  osTypeArray,
-  OsTypeArray,
   buildExceptionFilter,
 } from '../../lists/common';
