+++ x-pack/plugins/security_solution/public/common/components/autocomplete/field_value_match.tsx
@@ -15,10 +15,11 @@ import {
 } from '@elastic/eui';
 import { uniq } from 'lodash';
 
+import { ListOperatorTypeEnum as OperatorTypeEnum } from '@kbn/securitysolution-io-ts-list-types';
 import { IFieldType, IIndexPattern } from '../../../../../../../src/plugins/data/common';
 import { useFieldValueAutocomplete } from './hooks/use_field_value_autocomplete';
 import { paramIsValid, getGenericComboBoxProps } from './helpers';
-import { OperatorTypeEnum } from '../../../lists_plugin_deps';
+
 import { GetGenericComboBoxPropsReturn } from './types';
 import * as i18n from './translations';
