+++ x-pack/plugins/security_solution/public/common/components/autocomplete/field_value_match_any.tsx
@@ -9,11 +9,12 @@ import React, { useState, useCallback, useMemo } from 'react';
 import { EuiFormRow, EuiComboBoxOptionOption, EuiComboBox } from '@elastic/eui';
 import { uniq } from 'lodash';
 
+import { ListOperatorTypeEnum as OperatorTypeEnum } from '@kbn/securitysolution-io-ts-list-types';
 import { IFieldType, IIndexPattern } from '../../../../../../../src/plugins/data/common';
 import { useFieldValueAutocomplete } from './hooks/use_field_value_autocomplete';
 import { getGenericComboBoxProps, paramIsValid } from './helpers';
-import { OperatorTypeEnum } from '../../../lists_plugin_deps';
 import { GetGenericComboBoxPropsReturn } from './types';
+
 import * as i18n from './translations';
 
 interface AutocompleteFieldMatchAnyProps {
