+++ x-pack/plugins/security_solution/public/common/components/autocomplete/helpers.ts
@@ -8,6 +8,8 @@
 import dateMath from '@elastic/datemath';
 import { EuiComboBoxOptionOption } from '@elastic/eui';
 
+import type { Type } from '@kbn/securitysolution-io-ts-list-types';
+import type { ListSchema } from '../../../lists_plugin_deps';
 import { IFieldType } from '../../../../../../../src/plugins/data/common';
 
 import {
@@ -19,7 +21,6 @@ import {
 } from './operators';
 import { GetGenericComboBoxPropsReturn, OperatorOption } from './types';
 import * as i18n from './translations';
-import { ListSchema, Type } from '../../../lists_plugin_deps';
 
 /**
  * Returns the appropriate operators given a field type
