+++ x-pack/plugins/security_solution/public/common/components/autocomplete/hooks/use_field_value_autocomplete.ts
@@ -8,9 +8,9 @@
 import { useEffect, useState, useRef } from 'react';
 import { debounce } from 'lodash';
 
+import { ListOperatorTypeEnum as OperatorTypeEnum } from '@kbn/securitysolution-io-ts-list-types';
 import { IFieldType, IIndexPattern } from '../../../../../../../../src/plugins/data/common';
 import { useKibana } from '../../../../common/lib/kibana';
-import { OperatorTypeEnum } from '../../../../lists_plugin_deps';
 
 interface FuncArgs {
   fieldSelected: IFieldType | undefined;
