+++ x-pack/plugins/security_solution/public/common/components/autocomplete/types.ts
@@ -7,7 +7,10 @@
 
 import { EuiComboBoxOptionOption } from '@elastic/eui';
 
-import { OperatorEnum, OperatorTypeEnum } from '../../../lists_plugin_deps';
+import type {
+  ListOperatorEnum as OperatorEnum,
+  ListOperatorTypeEnum as OperatorTypeEnum,
+} from '@kbn/securitysolution-io-ts-list-types';
 
 export interface GetGenericComboBoxPropsReturn {
   comboOptions: EuiComboBoxOptionOption[];
