+++ x-pack/plugins/security_solution/public/common/components/exceptions/add_exception_modal/index.tsx
@@ -25,6 +25,7 @@ import {
   EuiComboBox,
   EuiComboBoxOptionOption,
 } from '@elastic/eui';
+import type { ExceptionListType } from '@kbn/securitysolution-io-ts-list-types';
 import {
   hasEqlSequenceQuery,
   isEqlRule,
@@ -34,9 +35,9 @@ import { Status } from '../../../../../common/detection_engine/schemas/common/sc
 import {
   ExceptionListItemSchema,
   CreateExceptionListItemSchema,
-  ExceptionListType,
   ExceptionBuilder,
 } from '../../../../../public/shared_imports';
+
 import * as i18nCommon from '../../../translations';
 import * as i18n from './translations';
 import * as sharedI18n from '../translations';
