+++ x-pack/plugins/security_solution/public/common/components/exceptions/use_fetch_or_create_rule_exception_list.test.tsx
@@ -12,7 +12,7 @@ import * as rulesApi from '../../../detections/containers/detection_engine/rules
 import * as listsApi from '../../../../../lists/public/exceptions/api';
 import { getExceptionListSchemaMock } from '../../../../../lists/common/schemas/response/exception_list_schema.mock';
 import { savedRuleMock } from '../../../detections/containers/detection_engine/rules/mock';
-import { ExceptionListType } from '../../../lists_plugin_deps';
+import type { ExceptionListType } from '@kbn/securitysolution-io-ts-list-types';
 import { ListArray } from '../../../../common/detection_engine/schemas/types';
 import { getListArrayMock } from '../../../../common/detection_engine/schemas/types/lists.mock';
 import {
@@ -20,6 +20,7 @@ import {
   UseFetchOrCreateRuleExceptionListProps,
   ReturnUseFetchOrCreateRuleExceptionList,
 } from './use_fetch_or_create_rule_exception_list';
+import { ExceptionListSchema } from '../../../shared_imports';
 
 const mockKibanaHttpService = coreMock.createStart().http;
 jest.mock('../../../detections/containers/detection_engine/rules/api');
@@ -31,7 +32,7 @@ describe('useFetchOrCreateRuleExceptionList', () => {
   let addEndpointExceptionList: jest.SpyInstance<
     ReturnType<typeof listsApi.addEndpointExceptionList>
   >;
-  let fetchExceptionListById: jest.SpyInstance<ReturnType<typeof listsApi.fetchExceptionListById>>;
+  let fetchExceptionListById: jest.SpyInstance<Promise<ExceptionListSchema>>;
   let render: (
     listType?: UseFetchOrCreateRuleExceptionListProps['exceptionListType']
   ) => RenderHookResult<
