+++ x-pack/plugins/security_solution/public/common/components/exceptions/viewer/helpers.tsx
@@ -7,8 +7,14 @@
 
 import moment from 'moment';
 
-import { entriesNested, ExceptionListItemSchema } from '../../../../lists_plugin_deps';
-import { getEntryValue, getExceptionOperatorSelect, formatOperatingSystems } from '../helpers';
+import { entriesNested } from '@kbn/securitysolution-io-ts-list-types';
+import {
+  ExceptionListItemSchema,
+  getEntryValue,
+  getExceptionOperatorSelect,
+} from '../../../../lists_plugin_deps';
+
+import { formatOperatingSystems } from '../helpers';
 import { FormattedEntry, BuilderEntry, DescriptionListItem } from '../types';
 import * as i18n from '../translations';
