+++ x-pack/plugins/security_solution/public/common/components/exceptions/viewer/index.test.tsx
@@ -11,11 +11,9 @@ import { ThemeProvider } from 'styled-components';
 
 import { ExceptionsViewer } from './';
 import { useKibana } from '../../../../common/lib/kibana';
-import {
-  ExceptionListTypeEnum,
-  useExceptionListItems,
-  useApi,
-} from '../../../../../public/lists_plugin_deps';
+import { useExceptionListItems, useApi } from '../../../../../public/lists_plugin_deps';
+
+import { ExceptionListTypeEnum } from '@kbn/securitysolution-io-ts-list-types';
 import { getExceptionListSchemaMock } from '../../../../../../lists/common/schemas/response/exception_list_schema.mock';
 import { getFoundExceptionListItemSchemaMock } from '../../../../../../lists/common/schemas/response/found_exception_list_item_schema.mock';
 import { getMockTheme } from '../../../lib/kibana/kibana_react.mock';
