+++ x-pack/plugins/security_solution/public/detections/components/value_lists_management_modal/form.tsx
@@ -18,7 +18,9 @@ import {
   EuiSelectOption,
 } from '@elastic/eui';
 
-import { useImportList, ListSchema, Type } from '../../../shared_imports';
+import type { Type } from '@kbn/securitysolution-io-ts-list-types';
+import { useImportList, ListSchema } from '../../../shared_imports';
+
 import * as i18n from './translations';
 import { useKibana } from '../../../common/lib/kibana';
