+++ x-pack/plugins/security_solution/public/detections/pages/detection_engine/rules/details/index.tsx
@@ -28,6 +28,7 @@ import { useDispatch } from 'react-redux';
 import styled from 'styled-components';
 import deepEqual from 'fast-deep-equal';
 
+import { ExceptionListTypeEnum } from '@kbn/securitysolution-io-ts-list-types';
 import {
   useDeepEqualSelector,
   useShallowEqualSelector,
@@ -83,7 +84,8 @@ import { ExceptionsViewer } from '../../../../../common/components/exceptions/vi
 import { DEFAULT_INDEX_PATTERN } from '../../../../../../common/constants';
 import { useGlobalFullScreen } from '../../../../../common/containers/use_full_screen';
 import { Display } from '../../../../../hosts/pages/display';
-import { ExceptionListTypeEnum, ExceptionListIdentifiers } from '../../../../../shared_imports';
+import type { ExceptionListIdentifiers } from '../../../../../shared_imports';
+
 import {
   focusUtilityBarAction,
   onTimelineTabKeyPressed,
