+++ x-pack/plugins/security_solution/public/shared_imports.ts
@@ -33,23 +33,23 @@ export { ERROR_CODE } from '../../../../src/plugins/es_ui_shared/static/forms/he
 
 export {
   exportList,
-  useIsMounted,
   useCursor,
   useApi,
   useAsync,
   useExceptionListItems,
   useExceptionLists,
-  usePersistExceptionItem,
-  usePersistExceptionList,
   useFindLists,
   useDeleteList,
   useImportList,
   useCreateListIndex,
   useReadListIndex,
   useReadListPrivileges,
-  addExceptionListItem,
-  updateExceptionListItem,
   fetchExceptionListById,
+  addIdToEntries,
+  getOperatorType,
+  getNewExceptionItem,
+  getEntryValue,
+  getExceptionOperatorSelect,
   addExceptionList,
   ExceptionListFilter,
   ExceptionListIdentifiers,
