+++ x-pack/plugins/security_solution/server/lib/detection_engine/signals/filters/create_field_and_set_tuples.test.ts
@@ -11,7 +11,7 @@ import { mockLogger, sampleDocWithSortId } from '../__mocks__/es_results';
 import { getExceptionListItemSchemaMock } from '../../../../../../lists/common/schemas/response/exception_list_item_schema.mock';
 import { listMock } from '../../../../../../lists/server/mocks';
 import { getSearchListItemResponseMock } from '../../../../../../lists/common/schemas/response/search_list_item_schema.mock';
-import { EntryList } from '../../../../../../lists/common';
+import { EntryList } from '@kbn/securitysolution-io-ts-list-types';
 import { buildRuleMessageMock as buildRuleMessage } from '../rule_messages.mock';
 
 describe('filterEventsAgainstList', () => {
